Add additional supported social media links
Description:
The social.md file lists several other options for social media that
were not included in the official documentation. I have added them
here for completeness (alternatively, a note could just be made in
the README which points to said file to see the most up-to-date list
of supported accounts).

Note: Documentation update